Changes UniqueMeshFunction to instantiate only on OWNED entities
This is required because UniqueMeshFunction is used on source terms,
and those source terms are only well posed on OWNED entities when they
are derived from subgrid meshes that are only instantiated on OWNED
entities.

This may break some things -- run tests and confirm please CI.  If it
doesn't break things, check with @lipnikov as to whether he is ok with
this change.